Revert "Merge master (#262)"
This reverts commit 0dc5962f28954f6a772f5c55ec301d27cdddd1bb.
