+++ .env.example
@@ -27,15 +27,13 @@ DATABASE_URL=mysql://root:root@localhost/arena
 # Used in mail templates
 ARENA_WEBSITE=http://localhost:8080
 
+# SMTP server address (smtp://user:password@host(:port)?)
 # You can use Nodemailer App (https://nodemailer.com/app/) or mailtrap.io to test the emails
-EMAIL_HOST=
-EMAIL_PORT=
-EMAIL_SECURE=
-EMAIL_SENDER_NAME=
-EMAIL_SENDER_ADDRESS=
-EMAIL_AUTH_USER=
-EMAIL_AUTH_PASSWORD=
-EMAIL_REJECT_UNAUTHORIZED=
+SMTP_URI=smtp://user:pass@address:25/?pool=true&maxConnections=1
+GMAIL=false
+GMAIL_USERNAME=uttarena@gmail.com
+GMAIL_PASSWORD=
+MAX_MAILS_PER_BATCH=100
 
 # Used to give a discount on tickets
 PARTNER_MAILS=utt.fr,utc.fr,utbm.fr
