+++ src/services/email/components.ts
@@ -1,5 +1,5 @@
 import { escape } from 'mustache';
-import type { Component, MailButton, MailTable } from './types';
+import type { Component } from './types';
 
 export const style = {
   text: {
@@ -35,7 +35,7 @@ export const escapeText = (text: string) =>
     .replaceAll(/_([^<>_]+)_/gi, '<i>$1</i>')
     .replaceAll(/\*([^*<>]+)\*/gi, '<strong>$1</strong>');
 
-const inflateButton = (item: MailButton) =>
+const inflateButton = (item: Component.Button) =>
   `<a target="_blank" href="${escapeText(
     item.location,
   )}" style="border:none;text-decoration:none;user-select:none"><button style="background-color:${
@@ -44,12 +44,12 @@ const inflateButton = (item: MailButton) =>
     style.text.font
   };outline:none;border:none;color:${style.button.font};cursor:pointer">${item.name}</button></a>`;
 
-const inflateButtonWrapper = (item: MailButton | MailButton[]) =>
+const inflateButtonWrapper = (item: Component.Button | Component.Button[]) =>
   `<tr><td><table style="border:none;border-spacing:5px"><tbody><tr>${
     Array.isArray(item) ? item.map(inflateButton).join('') : inflateButton(item)
   }</tr></tbody></table></td></tr>`;
 
-const inflateTable = (item: MailTable) => {
+const inflateTable = (item: Component.Table) => {
   const properties = Object.keys(item.items[0] ?? {});
   if (properties.length === 0 || item.items.length < 2) return '';
   return `${
@@ -99,8 +99,9 @@ const inflateText = (item: string) =>
 export const inflate = (content: Component): string => {
   if (typeof content === 'string') return inflateText(content);
   if (Array.isArray(content)) {
-    if (content.some((item: string | MailButton) => typeof item !== 'object')) return inflateList(<string[]>content);
-    return inflateButtonWrapper(<MailButton[]>content);
+    if (content.some((item: string | Component.Button) => typeof item !== 'object'))
+      return inflateList(<string[]>content);
+    return inflateButtonWrapper(<Component.Button[]>content);
   }
   if ('location' in content) return inflateButtonWrapper(content);
   if ('items' in content) return inflateTable(content);
